Add a generator that draws integers following a Zipfian distribution.
Although the WeightedRangeGenerator works well for selecting negative
samples, more than 1/5th of the training time is spend in this
iterator with large vocabulary sizes. Even though the number of
comparisons in the WeightedRangeGenerator is relatively small (due to
binary search), the prefix sum array does not fit CPU cache well,
and becomes memory latency-bound.

This generator approximates the token frequencies using a Zipfian
distribution with exponent 1. This generator barely shows up in the
profiles and makes training a fair bit faster.

Future work: the fastText paper suggests using the square root of a
token's frequency as the weight in negative sampling. Using this
Zipfian distribution gives a larger weight to more frequent
tokens. This may need more tweaking in the future.
