+++ finalfrontier/Cargo.toml
@@ -9,6 +9,7 @@ fnv = "1"
 ndarray = "0.11"
 ndarray-rand = "0.7"
 rand = "0.4"
+zipf = "3"
 
 [dev-dependencies]
 lazy_static = "1"
