+++ finalfrontier/src/lib.rs
@@ -19,11 +19,16 @@ extern crate ndarray_rand;
 
 extern crate rand;
 
+extern crate zipf;
+
 mod io;
 pub use io::SentenceIterator;
 
 mod config;
 pub use config::{Config, LossType, ModelType};
+
+mod sampling;
+pub use sampling::{WeightedRangeGenerator, ZipfRangeGenerator};
 
 mod subword;
 pub use subword::{NGrams, SubwordIndices};
